requirements.txt
kagglehub
pandas
snowflake-connector-python
pyarrow